Dynamically load PDF stuff into components
 + Use React.lazy to load components, and
 + Use plain dynamic import to load objects (combination of components
 and plain JS objects).

Also refactor `Download this analysis` into its own component so it
doesn't interfere we the remainder of Analysis component
