+++ takwimu/takwimu_ui/src/components/AnalysisContent/index.js
@@ -2,19 +2,9 @@
 import React, { Fragment, useEffect, useState } from 'react';
 import { PropTypes } from 'prop-types';
 
-import ReactPDF, {
-  Document,
-  Page,
-  Text,
-  View,
-  Image,
-  Link,
-  StyleSheet
-} from '@react-pdf/renderer';
-
 import { Typography, withStyles, Grid } from '@material-ui/core';
 
-import { countrify, RichTypography } from '../core';
+import { RichTypography } from '../core';
 import Actions from './Actions';
 import { Analysis as AnalysisReadNext } from '../Next';
 import CarouselTopic from './topics/CarouselTopic';
@@ -25,7 +15,6 @@ import RelatedContent from '../RelatedContent';
 import OtherInfoNav from './OtherInfoNav';
 
 import profileHeroImage from '../../assets/images/profile-hero-line.png';
-import logoWhite from '../../assets/images/logo-white-all.png';
 
 const styles = theme => ({
   root: {
@@ -63,182 +52,10 @@ const styles = theme => ({
   }
 });
 
-// Create styles
-const pdfStyles = StyleSheet.create({
-  page: {
-    padding: 20,
-    paddingBottom: 50
-  },
-  section: {
-    padding: 20
-  },
-  footer: {
-    position: 'absolute',
-    height: 50,
-    bottom: 0,
-    right: 40,
-    left: 40
-  },
-  header: {
-    position: 'relative',
-    height: 100,
-    marginLeft: -20,
-    marginRight: -20,
-    marginBottom: 20
-  },
-  divider: {
-    height: 4,
-    marginTop: 46,
-    backgroundColor: '#29a87c',
-    width: '100%'
-  },
-  downloadedAt: {
-    position: 'absolute',
-    fontSize: 14,
-    left: 0
-  },
-  linkTakwimuFooter: {
-    position: 'absolute',
-    fontSize: 14,
-    right: 0,
-    color: 'black',
-    textDecoration: 'none'
-  },
-  linkTakwimu: {
-    position: 'absolute',
-    top: 26,
-    fontSize: 14,
-    right: 20,
-    color: 'black',
-    textDecoration: 'none'
-  },
-  linkLicense: {
-    position: 'absolute',
-    top: 54,
-    fontSize: 14,
-    right: 20,
-    color: 'black',
-    textDecoration: 'none'
-  },
-  logo: {
-    width: 80,
-    marginTop: 42,
-    marginLeft: 10
-  },
-  logoBackground: {
-    position: 'absolute',
-    left: 40,
-    backgroundColor: '#29a87c',
-    width: 100,
-    height: 100,
-    borderRadius: 50,
-    borderBottomColor: '#ffffff',
-    borderBottomStyle: 'solid',
-    borderBottomWidth: 2
-  },
-  title: {
-    fontSize: 54
-    // fontFamily: 'Lora'
-  },
-  text: {
-    fontSize: 14,
-    lineHeight: 2.05,
-    // fontFamily: 'Muli',
-    paddingBottom: 28
-  },
-  boldText: {
-    fontSize: 14,
-    fontWeight: 500,
-    lineHeight: 2.05,
-    // fontFamily: 'Muli',
-    paddingBottom: 28
-  }
-});
-
-const AnalysisPDF = ({ data, topic, takwimu }) => (
-  <Document>
-    <Page size="A4" style={pdfStyles.page}>
-      <View style={pdfStyles.header} fixed>
-        <Link href="https://takwimu.africa" style={pdfStyles.linkTakwimu}>
-          www.takwimu.africa
-        </Link>
-        <View style={pdfStyles.divider} />
-        <Link
-          style={pdfStyles.linkLicense}
-          href="//creativecommons.org/licenses/by/4.0/"
-        >
-          2018 Takwimu CC by 4.0
-        </Link>
-        <View style={pdfStyles.logoBackground}>
-          <Image style={pdfStyles.logo} src={logoWhite} />
-        </View>
-      </View>
-      <View style={pdfStyles.section}>
-        <Text style={pdfStyles.title}>
-          {countrify(
-            data.content.title,
-            takwimu.country,
-            takwimu.countries,
-            ' : '
-          )}
-        </Text>
-      </View>
-      {topic === 'topic' ? (
-        <View style={pdfStyles.section}>
-          {data.content.body.map(c => {
-            if (c.type === 'text') {
-              return c.value
-                .split('</p>')
-                .map(t => (
-                  <Text style={pdfStyles.text}>
-                    {t.replace(/<(?:.|\n)*?>/gi, '')}
-                  </Text>
-                ));
-            }
-            return null;
-          })}
-        </View>
-      ) : (
-        <View style={pdfStyles.section}>
-          <Text style={pdfStyles.boldText}>
-            {data.item.name}, {data.item.title}
-          </Text>
-          {data.item.description.split('</p>').map(t => (
-            <Text style={pdfStyles.text}>
-              {t.replace(/<(?:.|\n)*?>/gi, '')}
-            </Text>
-          ))}
-        </View>
-      )}
-
-      <View style={pdfStyles.footer} fixed>
-        <Text style={pdfStyles.downloadedAt}>
-          Dowloaded{' '}
-          {new Date().toLocaleDateString('UTC', {
-            day: '2-digit',
-            month: 'short',
-            year: 'numeric'
-          })}
-        </Text>
-        <Link href="https://takwimu.africa" style={pdfStyles.linkTakwimuFooter}>
-          www.takwimu.africa
-        </Link>
-      </View>
-    </Page>
-  </Document>
-);
-
-AnalysisPDF.propTypes = {
-  data: PropTypes.shape({}).isRequired,
-  topic: PropTypes.oneOf(['topic', 'carousel_topic']).isRequired,
-  takwimu: PropTypes.shape({}).isRequired
-};
-
 function AnalysisContent({ classes, content, topicIndex, takwimu, onChange }) {
   const [carouselItemIndex, setCarouselItemIndex] = useState(
     content.body[topicIndex].type === 'carousel_topic' ? 0 : -1
   );
-  const [analysisBlob, setAnalysisBlob] = useState(null);
   const [id, setId] = useState(`${content.id}-${topicIndex}`);
 
   useEffect(() => {
@@ -260,22 +77,6 @@ function AnalysisContent({ classes, content, topicIndex, takwimu, onChange }) {
       script.src = 'https://public.flourish.studio/resources/embed.js';
       document.body.appendChild(script);
     }
-
-    ReactPDF.pdf(
-      <AnalysisPDF
-        topic={content.body[topicIndex].type}
-        data={{
-          item:
-            carouselItemIndex !== -1
-              ? content.body[topicIndex].value.body[carouselItemIndex]
-              : null,
-          content: content.body[topicIndex].value
-        }}
-        takwimu={takwimu}
-      />
-    )
-      .toBlob()
-      .then(setAnalysisBlob);
   }, [id, carouselItemIndex]);
 
   const showContent = index => () => {
@@ -286,6 +87,15 @@ function AnalysisContent({ classes, content, topicIndex, takwimu, onChange }) {
     profile_navigation: { value: profileNavigation },
     read_next: { value: readNext }
   } = content;
+
+  const topic = content.body[topicIndex].type;
+  const data = {
+    item:
+      carouselItemIndex !== -1
+        ? content.body[topicIndex].value.body[carouselItemIndex]
+        : null,
+    content: content.body[topicIndex].value
+  };
 
   return (
     <Fragment>
@@ -314,7 +124,9 @@ function AnalysisContent({ classes, content, topicIndex, takwimu, onChange }) {
         <Actions
           title={content.body[topicIndex].value.title}
           page={takwimu.page}
-          pdfBlob={analysisBlob}
+          topic={topic}
+          data={data}
+          takwimu={takwimu}
         />
 
         {content.body[topicIndex].type === 'carousel_topic' ? (
@@ -344,7 +156,14 @@ function AnalysisContent({ classes, content, topicIndex, takwimu, onChange }) {
           </Grid>
         )}
 
-        <Actions page={takwimu.page} pdfBlob={analysisBlob} hideLastUpdated />
+        <Actions
+          title={content.body[topicIndex].value.title}
+          page={takwimu.page}
+          topic={topic}
+          data={data}
+          takwimu={takwimu}
+          hideLastUpdated
+        />
         <ContentNavigation
           labelText={profileNavigation.title}
           labelTextStrong={content.title}
